# spec/openapi.yaml
openapi: 3.1.0
info:
  title: MFC Overlay Relay API
  version: 0.1.0
servers:
  - url: https://nowplaying.mfc.example
paths:
...
  /api/channels:
    get:
      summary: List channels
      security:
        - adminBearer: []
      responses:
        "200":
          description: Channel list without tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelListResponse"
    post:
      summary: Create a channel
      security:
        - adminBearer: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelRequest"
      responses:
        "201":
          description: Channel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelCreateResponse"
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
    publishBearer:
      type: http
      scheme: bearer
    settingsBearer:
      type: http
      scheme: bearer
...
  schemas:
    CreateChannelRequest:
      type: object
      properties:
        id:
          type: string
          pattern: "^[a-z0-9_-]{1,64}$"
        display_name:
          type: string
        settings:
          $ref: "#/components/schemas/SettingsPatch"
    ChannelCreateResponse:
      allOf:
        - $ref: "#/components/schemas/PublicChannel"
        - type: object
          properties:
            publish_token:
              type: string
            settings_token:
              type: string
    ChannelListResponse:
      type: object
      properties:
        channels:
          type: array
          items:
            $ref: "#/components/schemas/ChannelSummary"
    ChannelSummary:
      type: object
      properties:
        id:
          type: string
        display_name:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
        urls:
          type: object
          additionalProperties:
            type: string
...
    PublicChannel:
      type: object
      properties:
        id:
          type: string
        settings:
          $ref: "#/components/schemas/Settings"
        urls:
          type: object
          additionalProperties:
            type: string
...
    Settings:
      type: object
      properties:
        display_name:
          type: string
        template:
          type: string
          enum: [glassmorphic, compact-bar, minimal-clean, neon-cyber, spotify-dark]
        preset:
          type: string
          enum: [glass-pop, bubblegum, after-dark, cyber-candy, clean-luxe]
        accent_hex:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        accent_palette:
          type: array
          maxItems: 8
          items:
            type: string
            pattern: "^#[0-9a-fA-F]{6}$"
        color_mode:
          type: string
          enum: [solid, rotating]
        color_rotate_ms:
          type: integer
          minimum: 1000
          maximum: 30000
        glow_opacity:
          type: number
        card_opacity:
          type: number
        blur_radius:
          type: integer
        corner_radius:
          type: integer
        scale:
          type: number
        width_px:
          type: integer
        anchor:
          type: string
          enum: [top-left, top-center, top-right, center-left, center, center-right, bottom-left, bottom-center, bottom-right]
        offset_x_px:
          type: integer
        offset_y_px:
          type: integer
        show_paused:
          type: boolean
        stale_after_ms:
          type: integer
        custom_gif_enabled:
          type: boolean
        custom_gif_url:
          type: string
        social_rotation_enabled:
          type: boolean
        social_items:
          type: array
          maxItems: 12
          items:
            $ref: "#/components/schemas/SocialItem"
        album_rotation_enabled:
          type: boolean
        album_items:
          type: array
          maxItems: 12
          items:
            $ref: "#/components/schemas/AlbumItem"
        notice_enabled:
          type: boolean
        notice_items:
          type: array
          maxItems: 8
          items:
            $ref: "#/components/schemas/NoticeItem"
        ad_rotation_enabled:
          type: boolean
        ad_items:
          type: array
          maxItems: 12
          items:
            $ref: "#/components/schemas/AdItem"
        tile_rotate_ms:
          type: integer
          minimum: 2000
          maximum: 60000
        tile_rotation_order:
          type: array
          items:
            type: string
            enum: [social, album, ad, notice, now-playing]
        tile_size:
          type: string
          enum: [compact, large]
    SettingsPatch:
      allOf:
        - $ref: "#/components/schemas/Settings"
    SocialItem:
      type: object
      properties:
        label:
          type: string
        value:
          type: string
        url:
          type: string
        image_url:
          type: string
    AlbumItem:
      type: object
      properties:
        title:
          type: string
        caption:
          type: string
        image_url:
          type: string
        url:
          type: string
        published_at:
          type: string
    NoticeItem:
      type: object
      properties:
        message:
          type: string
        variant:
          type: string
          enum: [tip, promo, soft, hot]
    AdItem:
      type: object
      properties:
        title:
          type: string
        caption:
          type: string
        media_url:
          type: string
        media_type:
          type: string
          enum: [image, video]
        url:
          type: string
...